feat(asset-back): move inputs out of AssetImporterContext
Now ImporterFileInput and ImporterMemoryInput have their own .hpp file.
